+++ tensorpipe/channel/basic/context.h
@@ -11,8 +11,7 @@
 #include <memory>
 #include <string>
 
-#include <tensorpipe/channel/context.h>
-#include <tensorpipe/common/buffer.h>
+#include <tensorpipe/channel/cpu_context.h>
 #include <tensorpipe/common/callback.h>
 
 namespace tensorpipe {
