+++ tensorpipe/channel/cuda_ipc/channel.h
@@ -12,7 +12,7 @@
 
 #include <cuda_runtime.h>
 
-#include <tensorpipe/channel/channel.h>
+#include <tensorpipe/channel/cuda_context.h>
 #include <tensorpipe/channel/cuda_ipc/context.h>
 
 namespace tensorpipe {
