+++ tensorpipe/core/message.h
@@ -13,7 +13,7 @@
 #include <string>
 #include <vector>
 
-#include <tensorpipe/common/buffer.h>
+#include <tensorpipe/core/buffer.h>
 
 namespace tensorpipe {
